moved ams test client to its own file and added vars to represent each function
